+++ subsys/bluetooth/controller/coex/coex_ticker.c
@@ -13,7 +13,7 @@
 #include <zephyr/logging/log.h>
 #include <zephyr/drivers/gpio.h>
 #include <zephyr/sys/__assert.h>
-#include <zephyr/bluetooth/hci.h>
+#include <zephyr/bluetooth/hci_types.h>
 
 #include "controller/hal/ticker.h"
 #include "controller/ticker/ticker.h"
