+++ subsys/bluetooth/controller/ll_sw/isoal.c
@@ -14,7 +14,7 @@
 #include <zephyr/kernel.h>
 
 #include <zephyr/bluetooth/bluetooth.h>
-#include <zephyr/bluetooth/hci.h>
+#include <zephyr/bluetooth/hci_types.h>
 
 #include "util/memq.h"
