+++ subsys/bluetooth/controller/ll_sw/ll_addr.c
@@ -9,7 +9,7 @@
 
 #include <zephyr/kernel.h>
 #include <soc.h>
-#include <zephyr/bluetooth/hci.h>
+#include <zephyr/bluetooth/hci_types.h>
 #include <zephyr/bluetooth/controller.h>
 
 #include "hal/ccm.h"
