+++ subsys/bluetooth/controller/ll_sw/ll_feat.c
@@ -26,7 +26,7 @@
 #include "ll_feat.h"
 #include "ll_settings.h"
 
-#include <zephyr/bluetooth/hci.h>
+#include <zephyr/bluetooth/hci_types.h>
 
 #include "hal/debug.h"
