+++ subsys/bluetooth/controller/ll_sw/ll_tx_pwr.c
@@ -7,7 +7,7 @@
 
 #include <zephyr/kernel.h>
 #include <soc.h>
-#include <zephyr/bluetooth/hci.h>
+#include <zephyr/bluetooth/hci_types.h>
 #include <zephyr/bluetooth/hci_vs.h>
 
 #include "hal/cpu.h"
