+++ subsys/bluetooth/controller/ll_sw/nordic/lll/lll_adv_aux.c
@@ -8,7 +8,7 @@
 #include <stddef.h>
 
 #include <zephyr/sys/byteorder.h>
-#include <zephyr/bluetooth/hci.h>
+#include <zephyr/bluetooth/hci_types.h>
 #include <soc.h>
 
 #include "hal/cpu.h"
