+++ subsys/bluetooth/controller/ll_sw/nordic/lll/lll_scan.c
@@ -11,7 +11,7 @@
 #include <soc.h>
 
 #include <zephyr/sys/byteorder.h>
-#include <zephyr/bluetooth/hci.h>
+#include <zephyr/bluetooth/hci_types.h>
 
 #include "hal/cpu.h"
 #include "hal/ccm.h"
