+++ subsys/bluetooth/controller/ll_sw/nordic/lll/lll_test.c
@@ -41,7 +41,7 @@
 #include "ull_df_internal.h"
 #endif /* CONFIG_BT_CTLR_DTM_HCI_DF_IQ_REPORT */
 
-#include <zephyr/bluetooth/hci.h>
+#include <zephyr/bluetooth/hci_types.h>
 
 #include "hal/debug.h"
