+++ subsys/bluetooth/controller/ll_sw/nordic/ull/ull_iso_vendor.c
@@ -7,7 +7,7 @@
 #include <stdint.h>
 #include <stdbool.h>
 
-#include <zephyr/bluetooth/hci.h>
+#include <zephyr/bluetooth/hci_types.h>
 #include "isoal.h"
 #include "ull_iso_types.h"
