+++ subsys/bluetooth/controller/ll_sw/openisa/hal/RV32M1/ecb.c
@@ -8,7 +8,7 @@
 
 #include <string.h>
 
-#include <zephyr/bluetooth/hci.h>
+#include <zephyr/bluetooth/hci_types.h>
 
 #include <zephyr/sys/dlist.h>
 #include <zephyr/sys/byteorder.h>
