+++ subsys/bluetooth/controller/ll_sw/openisa/lll/lll_adv.c
@@ -9,7 +9,7 @@
 
 #include <zephyr/toolchain.h>
 #include <soc.h>
-#include <zephyr/bluetooth/hci.h>
+#include <zephyr/bluetooth/hci_types.h>
 #include <zephyr/sys/byteorder.h>
 
 #include "hal/cpu.h"
