+++ subsys/bluetooth/controller/ll_sw/ull_central_iso.c
@@ -49,7 +49,7 @@
 #include "ll.h"
 #include "ll_feat.h"
 
-#include <zephyr/bluetooth/hci.h>
+#include <zephyr/bluetooth/hci_types.h>
 
 #include "hal/debug.h"
