+++ subsys/bluetooth/controller/ll_sw/ull_df.c
@@ -8,7 +8,7 @@
 #include <zephyr/kernel.h>
 #include <soc.h>
 #include <zephyr/sys/util.h>
-#include <zephyr/bluetooth/hci.h>
+#include <zephyr/bluetooth/hci_types.h>
 
 #include "hal/cpu.h"
 #include "hal/ccm.h"
