+++ subsys/bluetooth/controller/ll_sw/ull_llcp_enc.c
@@ -6,7 +6,7 @@
 
 #include <zephyr/types.h>
 
-#include <zephyr/bluetooth/hci.h>
+#include <zephyr/bluetooth/hci_types.h>
 #include <zephyr/sys/byteorder.h>
 #include <zephyr/sys/slist.h>
 #include <zephyr/sys/util.h>
