+++ subsys/bluetooth/controller/ll_sw/ull_sched.c
@@ -5,7 +5,7 @@
  */
 
 #include <zephyr/sys/byteorder.h>
-#include <zephyr/bluetooth/hci.h>
+#include <zephyr/bluetooth/hci_types.h>
 
 #include "hal/ccm.h"
 #include "hal/radio.h"
